utils/ksampler_processor.py
from PIL import Image

class KSamplerProcessor:
    """Handles KSampler model processing."""
    
    @staticmethod
    def apply_ksampler(merged_img: Image, mask: Image, styled_outfit: Image) -> Image:
        """Pass everything through KSampler with Flux Fill Dev."""
        # TODO: Implement KSampler inference
        return merged_img