requirements.txt
# Core Dependencies
fastapi[all]

# Image Processing
Pillow==10.0.0
opencv-python==4.7.0.72

# Machine Learning Models
segment-anything==1.0

# Utilities
pydantic==1.10.7
python-multipart==0.0.6
aiofiles==23.1.0
python-dotenv==1.0.0
fastapi==0.99.1
fastapi-cli==0.0.7

# Grounding dino
numpy==1.26.4
supervision
wheel